Change count variable from int to NSUInteger
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/apps/gworkspace/trunk@35980 72102866-910b-0410-8b05-ffd578937521

+++ GWorkspace/Desktop/Dock/DockIcon.m
@@ -1,6 +1,6 @@
 /* DockIcon.m
  *  
- * Copyright (C) 2005-2012 Free Software Foundation, Inc.
+ * Copyright (C) 2005-2013 Free Software Foundation, Inc.
  *
  * Author: Enrico Sersale <enrico@imago.ro>
  * Date: January 2005
@@ -111,7 +111,7 @@
 
     if (isTrashIcon) {
       NSArray *subNodes;
-      int i, count;
+      NSUInteger i, count;
 
       ASSIGN (icon, [fsnodeRep trashIconOfSize: ceil(icnBounds.size.width)]);
       ASSIGN (trashFullIcon, [fsnodeRep trashFullIconOfSize: ceil(icnBounds.size.width)]);
@@ -533,7 +533,7 @@ x += 6; \
 
 - (void)setDraggedPaths:(NSArray *)paths
 {
-  int i;
+  NSUInteger i;
   
   [self unselect];
